Add front end to perform CRUD on Dog model. README updates.

+++ api-server/app/controllers/dogs_controller.rb
@@ -4,11 +4,18 @@ class DogsController < ApplicationController
   protect_from_forgery with: :null_session
 
   def index
-    dogs = Dog.all
+    dogs = Dog.all # TODO: paginate
+    byebug
     render json: dogs.as_json
   end
 
   def create
+    # reject unless bearer token matches ENV['API_KEY']
+    unless headers['Authorization'] == "Bearer #{ENV['API_KEY']}"
+      render json: { message: 'Unauthorized' }, status: :unauthorized
+      return
+    end
+
     dog = Dog.new(
       name: params[:name],
       description: params[:description]
@@ -31,6 +38,12 @@ class DogsController < ApplicationController
   end
 
   def destroy
+    # reject unless bearer token matches ENV['API_KEY']
+    unless headers['Authorization'] == "Bearer #{ENV['API_KEY']}"
+      render json: { message: 'Unauthorized' }, status: :unauthorized
+      return
+    end
+
     dog = Dog.find_by(id: params[:id])
     dog.destroy
     render json: { message: "Dog successfully destroyed. 🥲" }
